requirements.txt
aiohttp==3.8.4
aioselenium==0.0.1
aiosignal==1.3.1
alembic==1.10.4
aspose-pdf==23.4.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
colorama==0.4.6
configobj==5.0.8
configparser==5.3.0
dnspython==2.3.0
email-validator==2.0.0.post2
etelemetry==0.3.0
exceptiongroup==1.1.1
filelock==3.12.0
fitz==0.0.1.dev2
Flask==2.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
looseversion==1.1.2
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pathlib==1.0.1
pdf2image==1.16.3
Pillow==9.5.0
ping3==4.0.4
prov==2.0.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.7.0
PyMuPDF==1.22.3
pyparsing==3.0.9
PyPDF2==3.0.1
PyPDF4==1.27.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pyxnat==1.5
rdflib==6.3.2
requests==2.30.0
scipy==1.10.1
selenium==4.9.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.11
tqdm==4.65.0
traits==6.3.2
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Wand==0.6.11
Werkzeug==2.3.0
win32-setctime==1.1.0
wsproto==1.2.0
WTForms==3.0.1
wtforms-html5==0.6.1
yarl==1.9.2
